don't display delete button for last power user
